requirements.txt
absl-py==2.1.0
altair==5.5.0
arch==7.2.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clickhouse-connect==0.8.15
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
huggingface-hub==0.29.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.27.1
ncps==1.0.1
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.1
prettytable==3.15.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pydeck==0.9.1
pyecharts==2.0.8
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyWavelets==1.8.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==70.2.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.42.2
streamlit-echarts==0.4.0
streamlit-plotly-events==0.0.6
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zstandard==0.23.0
